requirements.txt
streamlit
tensorflow           # or tensorflow-macos if targeting mac M1 locally; for CI Ubuntu use plain tensorflow
numpy
pillow
